requirements.txt
astroid==2.0.4
colorama==0.4.0
dj-database-url==0.5.0
Django==1.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.6.1
pylint==2.1.1
six==1.11.0
virtualenv==16.1.0
whitenoise==3.3.1
gunicorn==19.6.0
wrapt==1.10.11
